Fix activities not being submitted on localhost resolution failure
On some hosts with specific networking configurations, `InetAddress.getLocalHost().getHostName()` fails with an `UnknownHostException`.  This commit changes the behaviour of `ActivitiesRecorder` such that it does not fail completely in such a case, resulting in activities never being pushed, but instead causes it to send a hostname of `unknown` to the server.